+++ packages/bignumber/lib/_version.d.ts
@@ -1,2 +1,2 @@
-export declare const version = "bignumber/5.4.0";
+export declare const version = "bignumber/5.4.1";
 //# sourceMappingURL=_version.d.ts.map