+++ packages/networks/lib/index.js
@@ -21,17 +21,8 @@ function ethDefaultProvider(network) {
             catch (error) { }
         }
         if (providers.EtherscanProvider) {
-            //try {
-            //    providerList.push(new providers.EtherscanProvider(network, options.etherscan));
-            //} catch(error) { }
-            // These networks are currently faulty on this provider
-            // @TODO: This goes away once they have fixed their nodes
-            var skip = ["ropsten"];
             try {
-                var provider = new providers.EtherscanProvider(network);
-                if (provider.network && skip.indexOf(provider.network.name) === -1) {
-                    providerList.push(provider);
-                }
+                providerList.push(new providers.EtherscanProvider(network, options.etherscan));
             }
             catch (error) { }
         }