+++ packages/providers/lib.esm/base-provider.js
@@ -1065,6 +1065,9 @@ export class BaseProvider extends Provider {
             yield this.getNetwork();
             const hexTx = yield Promise.resolve(signedTransaction).then(t => hexlify(t));
             const tx = this.formatter.transaction(signedTransaction);
+            if (tx.confirmations == null) {
+                tx.confirmations = 0;
+            }
             const blockNumber = yield this._getInternalBlockNumber(100 + 2 * this.pollingInterval);
             try {
                 const hash = yield this.perform("sendTransaction", { signedTransaction: hexTx });
@@ -1087,7 +1090,7 @@ export class BaseProvider extends Provider {
                 }
                 tx[key] = Promise.resolve(values[key]).then((v) => (v ? this._getAddress(v) : null));
             });
-            ["gasLimit", "gasPrice", "value"].forEach((key) => {
+            ["gasLimit", "gasPrice", "maxFeePerGas", "maxPriorityFeePerGas", "value"].forEach((key) => {
                 if (values[key] == null) {
                     return;
                 }
@@ -1245,7 +1248,9 @@ export class BaseProvider extends Provider {
                             tx.confirmations = confirmations;
                         }
                     }
-                    return this.formatter.blockWithTransactions(block);
+                    const blockWithTxs = this.formatter.blockWithTransactions(block);
+                    blockWithTxs.transactions = block.transactions.map((tx) => this._wrapTransaction(tx));
+                    return blockWithTxs;
                 }
                 return this.formatter.block(block);
             }), { oncePoll: this });