+++ packages/tests/src.ts/test-providers.ts
@@ -312,7 +312,7 @@ const blockchainData: { [ network: string ]: TestCases } = {
                 type: 2,
                 creates: null,
                 from: '0xad252DD6C011E613610A36368f04aC84D5185b7c',
-                gasPrice: bnify("0x0268ab0ed6"),
+                //gasPrice: bnify("0x0268ab0ed6"),
                 maxPriorityFeePerGas: bnify("0x0268ab0ed6"),
                 maxFeePerGas: bnify("0x0268ab0ed6"),
                 gasLimit: bnify("0x5208"),
@@ -743,7 +743,9 @@ Object.keys(blockchainData).forEach((network) => {
 
             return tx;
         }, test, (provider: string, network: string, test: TestDescription) => {
-            return false;
+            // Temporary; pocket is being broken again for old transactions
+            return provider === "PocketProvider";
+            //return false;
         });
     });
 
@@ -763,19 +765,9 @@ Object.keys(blockchainData).forEach((network) => {
 
             return receipt;
         }, test, (provider: string, network: string, test: TestDescription) => {
-
-            // @TODO: Remove once Etherscan fixes whatever makes this unhappy
-            if (provider === "EtherscanProvider") {
-                if (hash === "0x55c477790b105e69e98afadf0505cbda606414b0187356137132bf24945016ce") {
-                    return true;
-                }
-
-                if (hash === "0xf724f1d6813f13fb523c5f6af6261d06d41138dd094fff723e09fb0f893f03e6") {
-                    return true;
-                }
-            }
-
-            return false;
+            // Temporary; pocket is being broken again for old transactions
+            return provider === "PocketProvider";
+            //return false;
         });
     });
 });
@@ -786,9 +778,7 @@ Object.keys(blockchainData).forEach((network) => {
             name: `throws correct ${ code } error`,
             networks: [ "ropsten" ],
             checkSkip: (provider: string, network: string, test: TestDescription) => {
-                // @TODO: Remove once Etherscan supports EIP-1559
-                return (code === ethers.utils.Logger.errors.UNPREDICTABLE_GAS_LIMIT && provider === "EtherscanProvider");
-                //return false;
+                return false;
             },
             execute: async (provider: ethers.providers.Provider) => {
                 try {